+++ packages/vitepress-plugin-pagefind/src/type.ts
@@ -147,6 +147,12 @@ export interface SearchConfig {
    * @default false
    */
   manual?: boolean
+
+  /**
+   * debouncedSearch delay
+   * @default 300
+   */
+  delay?: number
 }
 
 export type PagefindConfig = PagefindOption & SearchConfig